requirements.txt
langchain>=0.1.0
langchain-apify>=0.0.3
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.10
apify-client>=1.6.1
apify>=1.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
typing-extensions>=4.8.0
pytest>=7.4.0 